Updated to override the existing (non-rgraph) charts with the bodies of the new ones.
