--- modules/Charts/Dashlets/RGraph_OpportunitiesByLeadSourceDashlet/RGraph_OpportunitiesByLeadSourceConfigure.tpl
@@ -1,82 +1,0 @@
-{*
-
-/*********************************************************************************
- * SugarCRM Community Edition is a customer relationship management program developed by
- * SugarCRM, Inc. Copyright (C) 2004-2013 SugarCRM Inc.
-
- * SuiteCRM is an extension to SugarCRM Community Edition developed by Salesagility Ltd.
- * Copyright (C) 2011 - 2014 Salesagility Ltd.
- *
- * This program is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Affero General Public License version 3 as published by the
- * Free Software Foundation with the addition of the following permission added
- * to Section 15 as permitted in Section 7(a): FOR ANY PART OF THE COVERED WORK
- * IN WHICH THE COPYRIGHT IS OWNED BY SUGARCRM, SUGARCRM DISCLAIMS THE WARRANTY
- * OF NON INFRINGEMENT OF THIRD PARTY RIGHTS.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE.  See the GNU Affero General Public License for more
- * details.
- *
- * You should have received a copy of the GNU Affero General Public License along with
- * this program; if not, see http://www.gnu.org/licenses or write to the Free
- * Software Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA
- * 02110-1301 USA.
- *
- * You can contact SugarCRM, Inc. headquarters at 10050 North Wolfe Road,
- * SW2-130, Cupertino, CA 95014, USA. or at email address contact@sugarcrm.com.
- *
- * The interactive user interfaces in modified source and object code versions
- * of this program must display Appropriate Legal Notices, as required under
- * Section 5 of the GNU Affero General Public License version 3.
- *
- * In accordance with Section 7(b) of the GNU Affero General Public License version 3,
- * these Appropriate Legal Notices must retain the display of the "Powered by
- * SugarCRM" logo and "Supercharged by SuiteCRM" logo. If the display of the logos is not
- * reasonably feasible for  technical reasons, the Appropriate Legal Notices must
- * display the words  "Powered by SugarCRM" and "Supercharged by SuiteCRM".
- ********************************************************************************/
-
-
-
-
-*}
-
-
-<div style='width: 400px'>
-<form name='configure_{$id}' action="index.php" method="post" onSubmit='return SUGAR.dashlets.postForm("configure_{$id}", SUGAR.mySugar.uncoverPage);' />
-<input type='hidden' name='id' value='{$id}' />
-<input type='hidden' name='module' value='{$module}' />
-<input type='hidden' name='action' value='DynamicAction' />
-<input type='hidden' name='DynamicAction' value='configureDashlet' />
-<input type='hidden' name='to_pdf' value='true' />
-<input type='hidden' name='configure' value='true' />
-<input type='hidden' id='dashletType' name='dashletType' value='{$dashletType}' />
-
-<table width="400" cellpadding="10" cellspacing="0" border="0" class="edit view" align="center">
-<tr>
-    <td valign='top' nowrap class='dataLabel'>{$LBL_LEAD_SOURCES}</td>
-    <td valign='top' class='dataField'>
-    	<select name="pbls_lead_sources[]" multiple size='3'>
-    		{$selected_datax}
-    	</select>
-    </td>
-</tr>
-<tr>
-    <td valign='top' nowrap class='dataLabel'>{$LBL_USERS}</td>
-	<td valign='top' class='dataField'>
-		<select name="pbls_ids[]" multiple size='3'>
-			{$pbls_ids}
-		</select>
-	</td>
-</tr>
-
-<tr>
-    <td align="right" colspan="2">
-        <input type='submit' onclick="" class='button' value='{$LBL_SUBMIT_BUTTON_LABEL}'>
-   	</td>
-</tr>
-</table>
-</form>
-</div>
